requirements.txt
pandas>=1.3.0,<2.0.0
openpyxl>=3.0.0,<4.0.0
Pillow>=9.0.0,<11.0.0
python-dotenv>=0.19.0,<1.0.0
